+++ CHANGELOG.md
@@ -3,13 +3,17 @@
 ## Version 0.2.0 (Under development)
 
 - [REFACTORING] Adding the `FlowingMenuTransitionStatus` class to allow animation testing
+- [RENAMING] Renaming `flowingMenuTransitionManagerNeedsPresentMenu:` to `flowingMenuNeedsPresentMenu:`
+- [RENAMING] Renaming `flowingMenuTransitionManagerNeedsDismissMenu:` to `flowingMenuNeedsDismissMenu:`
+- [RENAMING] Renaming `colorOfElasticShapeInFlowingMenuTransitionManager:` to `colorOfElasticShapeInFlowingMenu:`
+- [RENAMING] Renaming `flowingMenuTransitionManager:widthOfMenuView:` to `flowingMenu:widthOfMenuView:`
 
 ## [Version 0.1.0](https://github.com/yannickl/FlowingMenu/releases/tag/0.1.0)
 Released on 2015-12-04.
 
 - `flowingMenuTransitionManager:widthOfMenuView:` delegate method to get the width menu
 - `colorOfElasticShapeInFlowingMenuTransitionManager:` delegate method to get the color of the elastic shape
-- `flowingMenuTransitionManagerNeedsPresentMenu:` and `flowingMenuTransitionManagerNeedsDismissMenu` delegate methods to respond to interactive event
+- `flowingMenuTransitionManagerNeedsPresentMenu:` and `flowingMenuTransitionManagerNeedsDismissMenu:` delegate methods to respond to interactive event
 - `setInteractivePresentationView:` and `setInteractiveDismissView:` methods define views attached with gestures
 - Cocoapods support
 - Carthage support